Simplified to plugin template
